add reader model and add navbar to html

+++ model/src/main/java/com/brest/library/Book.java
@@ -3,10 +3,7 @@ package com.brest.library;
 import lombok.Getter;
 import lombok.Setter;
 
-import javax.persistence.Entity;
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
-import javax.persistence.Id;
+import javax.persistence.*;
 
 
 @Entity
@@ -16,8 +13,10 @@ public class Book {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer id;
+    private Long id;
 
     private String author;
     private String name;
+
+
 }
